notify-service/requirements.txt
azure-cosmos
schedule
pytz
python-dateutil